Do not for using 64-bit integers

+++ algorithms/src/Kokkos_Random.hpp
@@ -587,11 +587,13 @@ struct Random_XorShift1024_State<false> {
                                             int state_idx)
       : state_(&v(state_idx, 0)), stride_(v.stride_1()) {}
 
+  // NOLINTBEGIN(bugprone-implicit-widening-of-multiplication-result)
   KOKKOS_FUNCTION
-  uint64_t operator[](size_t i) const { return state_[i * stride_]; }
+  uint64_t operator[](int i) const { return state_[i * stride_]; }
 
   KOKKOS_FUNCTION
-  uint64_t& operator[](size_t i) { return state_[i * stride_]; }
+  uint64_t& operator[](int i) { return state_[i * stride_]; }
+  // NOLINTEND(bugprone-implicit-widening-of-multiplication-result)
 };
 
 template <class ExecutionSpace>
